Use correct type in slots connected to buttons
The buttonClicked signal from a QButtonGroup uses int so let's use that
in the slot connected to it as well.

Fixes #29

+++ example/qatemswitcher/mainwindow.cpp
@@ -215,13 +215,13 @@ void MainWindow::onAtemConnected()
     }
 }
 
-void MainWindow::changeProgramInput(quint16 input)
+void MainWindow::changeProgramInput(int input)
 {
     QAtemMixEffect *me = m_atemConnection->mixEffect(0);
 
     if(me)
     {
-        me->changeProgramInput(input);
+        me->changeProgramInput(static_cast<quint16>(input));
     }
     else
     {
@@ -229,13 +229,13 @@ void MainWindow::changeProgramInput(quint16 input)
     }
 }
 
-void MainWindow::changePreviewInput(quint16 input)
+void MainWindow::changePreviewInput(int input)
 {
     QAtemMixEffect *me = m_atemConnection->mixEffect(0);
 
     if(me)
     {
-        me->changePreviewInput(input);
+        me->changePreviewInput(static_cast<quint16>(input));
     }
     else
     {
@@ -341,9 +341,9 @@ void MainWindow::setTransitionStyle(quint8 me, quint8 style)
     m_transitionStyleGroup->button(style)->setChecked(true);
 }
 
-void MainWindow::changeTransitionStyle(quint8 style)
+void MainWindow::changeTransitionStyle(int style)
 {
-    m_atemConnection->mixEffect(0)->setTransitionType(style);
+    m_atemConnection->mixEffect(0)->setTransitionType(static_cast<quint8>(style));
 }
 
 void MainWindow::updateKeysOnNextTransition(quint8 me, quint8 keyers)
