+++ example/qatemswitcher/mainwindow.h
@@ -23,8 +23,8 @@ protected slots:
 
     void onAtemConnected();
 
-    void changeProgramInput(quint16 input);
-    void changePreviewInput(quint16 input);
+    void changeProgramInput(int input);
+    void changePreviewInput(int input);
     void updateProgramInput(quint8 me, quint16 oldInput, quint16 newInput);
     void updatePreviewInput(quint8 me, quint16 oldInput, quint16 newInput);
 
@@ -39,7 +39,7 @@ protected slots:
 
     void setTransitionRate(quint8 me, quint8 rate);
     void setTransitionStyle(quint8 me, quint8 style);
-    void changeTransitionStyle(quint8 style);
+    void changeTransitionStyle(int style);
     void updateKeysOnNextTransition(quint8 me, quint8 keyers);
     void changeKeysTransition(int btn, bool state);
     void setTransitionPosition(quint8 me, quint16 pos);
